# packages/sdk/.openapi-cache.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Cepage",
    "version": "0.1.0",
    "description": "HTTP API exposing the Cepage typed-skill library: list catalog, run skills, manage schedules, and author new skills from sessions."
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "Relative path under the Cepage HTTP server."
    }
  ],
...
  "paths": {
...
    "/skill-compiler/dry-run": {
      "post": {
        "tags": [
          "skill-compiler"
        ],
        "summary": "Validate a skill against inputs without executing it.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DryRunBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Dry-run report.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DryRunResult"
                }
              }
            }
          },
          "404": {
            "description": "Skill not found."
          }
        }
      }
    },
    "/skill-compiler/sessions/{sessionId}/preview": {
      "get": {
        "tags": [
          "skill-compiler"
        ],
        "summary": "Preview a compilation without persisting.",
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "agentType",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "opencode",
                "cursor"
              ],
              "default": "opencode"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Compilation preview.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompilationPreview"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DryRunBody": {
        "type": "object",
        "required": [
          "skillId",
          "inputs"
        ],
        "properties": {
          "skillId": {
            "type": "string"
          },
          "inputs": {
            "type": "object",
            "additionalProperties": true
          },
          "mode": {
            "type": "string",
            "enum": [
              "strict",
              "permissive"
            ]
          }
        }
      },
      "DryRunResult": {
        "type": "object",
        "required": [
          "overall",
          "checks",
          "warnings",
          "errors",
          "estimatedCost"
        ],
        "properties": {
          "overall": {
            "type": "string",
            "enum": [
              "PASS",
              "FAIL"
            ]
          },
          "checks": {
            "type": "object",
            "required": [
              "parametric",
              "schema",
              "graph"
            ],
            "properties": {
              "parametric": {
                "type": "string",
                "enum": [
                  "PASS",
                  "FAIL"
                ]
              },
              "schema": {
                "type": "string",
                "enum": [
                  "PASS",
                  "FAIL"
                ]
              },
              "graph": {
                "type": "string",
                "enum": [
                  "PASS",
                  "FAIL"
                ]
              }
            }
          },
          "warnings": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "object",
              "required": [
                "check",
                "message"
              ],
              "properties": {
                "check": {
                  "type": "string"
                },
                "field": {
                  "type": "string",
                  "nullable": true
                },
                "message": {
                  "type": "string"
                }
              }
            }
          },
          "estimatedCost": {
            "type": "number"
          }
        }
      },
      "CompilationPreview": {
        "type": "object",
        "required": [
          "skill",
          "report"
        ],
        "properties": {
          "skill": {
            "type": "object",
            "additionalProperties": true
          },
          "report": {
            "type": "object",
            "required": [
              "parameters",
              "estimatedCost",
              "graphStats",
              "warnings"
            ],
            "properties": {
              "parameters": {
                "type": "array",
                "items": {
                  "type": "object",
                  "required": [
                    "name",
                    "occurrences",
                    "inferredType"
                  ],
                  "properties": {
                    "name": {
                      "type": "string"
                    },
                    "occurrences": {
                      "type": "integer"
                    },
                    "inferredType": {
                      "type": "string"
                    },
                    "hint": {
                      "type": "string",
                      "nullable": true
                    }
                  }
                }
              },
              "estimatedCost": {
                "type": "number"
              },
              "graphStats": {
                "type": "object",
                "required": [
                  "nodes",
                  "edges"
                ],
                "properties": {
                  "nodes": {
                    "type": "integer"
                  },
                  "edges": {
                    "type": "integer"
                  }
                }
              },
              "warnings": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Optional for now. In Phase 3 the `local-user` shim is retired and a real Bearer token is required."
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    },
    {}
  ]
}